refactor: 데이터 패키지 구조 변경
