+++ app/src/main/java/woowacourse/shopping/RepositoryProvider.kt
@@ -3,7 +3,7 @@ package woowacourse.shopping
 import android.content.Context
 import woowacourse.shopping.data.goods.repository.GoodsRepositoryImpl
 import woowacourse.shopping.data.recentgoods.repository.RecentGoodsRepositoryImpl
-import woowacourse.shopping.data.shoppingcart.database.ShoppingDatabase
+import woowacourse.shopping.data.database.ShoppingDatabase
 import woowacourse.shopping.data.shoppingcart.repository.ShoppingCartRepositoryImpl
 import woowacourse.shopping.domain.repository.GoodsRepository
 import woowacourse.shopping.domain.repository.RecentGoodsRepository
