+++ app/src/main/java/woowacourse/shopping/data/recentgoods/RecentGoodsDao.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.recentgoods.database
+package woowacourse.shopping.data.recentgoods
 
 import androidx.room.Dao
 import androidx.room.Insert
