+++ app/src/main/java/woowacourse/shopping/data/recentgoods/mapper/RecentGoodsMapper.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.data.recentgoods.mapper
 
-import woowacourse.shopping.data.recentgoods.database.RecentGoodsEntity
+import woowacourse.shopping.data.recentgoods.RecentGoodsEntity
 import woowacourse.shopping.domain.model.Goods
 
 fun Goods.toEntity(time: Long): RecentGoodsEntity {
